Done Issue #7
The only variation with the original description of the issue is that
there are 3 functions instead of 2. 2 of them are generic functions
that read all the grayscale and color images from a given directory.
The other one is just focused on the Kaggle directory. This last one
also returns the labels provided on the Kaggle train table.